Fixes
Tests remain to do
