+++ server/models/components/utils.py
@@ -1,21 +1,9 @@
 from scipy.optimize import fsolve
-import matplotlib.pyplot as plt
-
-
-def factorial(k) :
-    
-    fact = 1
-    
-    for i in range(1, k + 1) :
-        fact *= i
-    
-    return fact
-
-def binomial_coeff(k, n) :
-    
-    return factorial(n) / (factorial(k) * factorial(n - k))
+from scipy.special import binom
 
 def compute_param(dms, max_days, prob_type):
+    
+    assert prob_type in ['exponential', 'binomial'], "Choose a valid name for the type of residuals"
     
     if prob_type == 'exponential':
         def f(x):
@@ -23,27 +11,33 @@ def compute_param(dms, max_days, prob_type):
         solutions = fsolve(f, 5)
         param = solutions[0]
     elif prob_type == 'binomial':
-        param = (dms - 1) / max_days
+        param = (dms - 1) / (max_days - 1)
         
     return param
 
-def compute_khi(dms, max_days = 21, prob_type = 'binomial') :
+def compute_khi_binom(dms, max_days = 21) :
     
     khi_tab = []
-    param = compute_param(dms, max_days, prob_type)
+    param = compute_param(dms, max_days, 'binomial')
     
-    if prob_type == 'exponential':
-        for k in range(max_days - 1):
-            khi = param ** k * (1 - param)
-            khi_tab.append(khi)
-        khi_tab.append(param ** (max_days - 1))
-    elif prob_type == 'binomial':
-        for k in range(max_days):
-            khi = binomial_coeff(k, max_days) * param ** k * (1 - param) ** (max_days - k)
-            khi_tab.append(khi)
+    for k in range(max_days):
+        khi = binom(max_days - 1, k) * param ** k * (1 - param) ** (max_days - 1 - k)
+        khi_tab.append(khi)
     
     return khi_tab
+
+def compute_khi_exp(dms, max_days = 21):
     
+    khi_tab = []
+    param = compute_param(dms, max_days, 'exponential')
+    
+    for k in range(max_days - 1):
+        khi = param ** k * (1 - param)
+        khi_tab.append(khi)
+    khi_tab.append(param ** (max_days - 1))
+    
+    return khi_tab
+
 def compute_residuals(khi_tab):
     
     residuals = [1]
@@ -58,18 +52,3 @@ def compute_delay_ki(duration):
         return [1]
     else:
         return [0]*(duration-2) + [0.5, 0.5]
-
-
-def compute_area_and_expectation(khi_tab, residuals):
-    
-    area = 0
-    expectation_contribs = []
-    N = len(khi_tab)
-    
-    for k in range(N) :
-        area += (residuals[k] + residuals[k + 1]) / 2
-        expectation_contribs.append((k + 1) * khi_tab[k])
-    
-    expectation = sum(expectation_contribs)
-    
-    return area, expectation, sum(khi_tab)
